enable multi-gpu inference
