+++ examples/generative/corrdiff/conf/config_generate.yaml
@@ -18,9 +18,9 @@ hydra:
 
 
 ## Main options
-res_ckpt_filename: "./checkpoints/diffusion.pkl"   
+res_ckpt_filename: "./checkpoints/diffusion.pt"   
   # Checkpoint filename for the generative model  
-reg_ckpt_filename: "./checkpoints/regression.pkl"
+reg_ckpt_filename: "./checkpoints/regression.pt"
   # Checkpoint filename for the mean predictor model
 image_outdir: "image_outdir"
   # Where to save the output images
@@ -32,15 +32,15 @@ num_steps: 18
   # Number of sampling steps
 sample_res: "full"
   # Sampling resolution
-pretext: null
-  # String to include the task
 res_edm: true
   # If residual-based edm is used
 sampling_method: "stochastic"
   # Sampling method ["stochastic", "deterministic"]
+seed_batch_size: 4
+  # batch size across the seed dimension
 
 ## Data options
-train_data_path: "./train_data"
+train_data_path: "s3://sw_climate_fno/nbrenowitz/2023-01-24-cwb-4years.zarr"  #"./train_data"
 patch_size: 448
 crop_size_x: 448
 crop_size_y: 448
@@ -61,6 +61,9 @@ N_grid_channels: 4
 normalization: "v1"
 times:
   - 2021-02-02T00:00:00
+#  - 2021-02-07T07:00:00
+#  - 2021-02-08T16:00:00
+#  - 2021-02-09T20:00:00
 
 ## Deterministic sampler options
 sigma_min: null
