+++ modulus/launch/logging/launch.py
@@ -15,8 +15,6 @@
 import re
 import sys
 import time
-from os import getcwd, makedirs
-from os.path import abspath, exists, join
 from typing import Dict, Tuple, Union
 
 import torch
@@ -223,7 +221,7 @@ class LaunchLogger(object):
             self.epoch_losses[name] = process_loss
             # Compute global loss
             if DistributedManager.is_initialized() and DistributedManager().distributed:
-                self.epoch_losses[name] = gather_loss(process_loss)
+                self.epoch_losses[f"Global {name}"] = gather_loss(process_loss)
 
         if self.root:
             # Console printing
@@ -292,6 +290,7 @@ class LaunchLogger(object):
         self,
         metric_dict: Dict[str, float],
         step: Tuple[str, int] = None,
+        print: bool = False,
     ):
         """Logs a dictionary of metrics to different supported backends
 
@@ -324,51 +323,6 @@ class LaunchLogger(object):
             # different intervals
             metric_dict[step[0]] = step[1]
             wandb.log(metric_dict)
-
-    def log_figure(
-        self,
-        figure,
-        artifact_file: str = "artifact",
-        plot_dir: str = "./",
-        log_to_file: bool = False,
-    ):
-        """Logs figures on root process to wand or mlflow. Will store it to file in case neither are selected.
-
-        Parameters
-        ----------
-        figure : Figure
-            matplotlib or plotly figure to plot
-        artifact_file : str, optional
-            File name. CAUTION overrides old files of same name
-        plot_dir : str, optional
-            output directory for plot
-        log_to_file : bool, optional
-            set to true in case figure shall be stored to file in addition to logging it to mlflow/wandb
-        """
-        dist = DistributedManager()
-        if dist.rank != 0:
-            return
-
-        if self.wandb_backend:
-            wandb.log({artifact_file: figure})
-
-        if self.mlflow_backend:
-            self.mlflow_client.log_figure(
-                figure=figure,
-                artifact_file=artifact_file,
-                run_id=self.mlflow_run.info.run_id,
-            )
-
-        if (not self.wandb_backend) and (not self.mlflow_backend):
-            log_to_file = True
-
-        if log_to_file:
-            plot_dir = abspath(join(getcwd(), plot_dir))
-            if not exists(plot_dir):
-                makedirs(plot_dir)
-            if not artifact_file.endswith(".png"):
-                artifact_file += ".png"
-            figure.savefig(join(plot_dir, artifact_file))
 
     @classmethod
     def toggle_wandb(cls, value: bool):
@@ -407,19 +361,14 @@ class LaunchLogger(object):
             PythonLogger().warning("WandB not initialized, turning off")
             use_wandb = False
 
+        if LaunchLogger.mlflow_run is None and use_mlflow:
+            PythonLogger().warning("MLFlow not initialized, turning off")
+            use_mlflow = False
+
         if use_wandb:
             LaunchLogger.toggle_wandb(True)
             wandb.define_metric("epoch")
             wandb.define_metric("iter")
-
-        # let only root process log to mlflow
-        if DistributedManager.is_initialized():
-            if DistributedManager().rank != 0:
-                return
-
-        if LaunchLogger.mlflow_run is None and use_mlflow:
-            PythonLogger().warning("MLFlow not initialized, turning off")
-            use_mlflow = False
 
         if use_mlflow:
             LaunchLogger.toggle_mlflow(True)
