+++ modulus/launch/logging/mlflow.py
@@ -26,6 +26,7 @@ from modulus.distributed import DistributedManager
 
 from .console import PythonLogger
 from .launch import LaunchLogger
+from .utils import create_ddp_group_tag
 
 logger = PythonLogger("mlflow")
 
@@ -81,16 +82,16 @@ def initialize_mlflow(
         Returns MLFlow logging client and active run object
     """
     dist = DistributedManager()
-    if dist.rank != 0:  # only root process should be logging to mlflow
-        return
-
-    start_time = datetime.now().astimezone()
-    time_string = start_time.strftime("%m/%d/%y_%H-%M-%S")
-    group_name = f"{run_name}_{time_string}"
-
+    if DistributedManager.is_initialized() and dist.distributed:
+        group_name = create_ddp_group_tag(run_name)
+        run_name = f"{run_name}-Process_{dist.rank}"
+    else:
+        start_time = datetime.now().astimezone()
+        time_string = start_time.strftime("%m/%d/%y_%H-%M-%S")
+        group_name = f"{run_name}_{time_string}"
     # Set default value here for Hydra
     if tracking_location is None:
-        tracking_location = str(Path("./mlruns").absolute())
+        tracking_location = str(Path(f"./mlruns_{dist.rank}").absolute())
 
     # Set up URI (remote or local)
     if mode == "online":
