+++ modulus/models/afno/distributed/afno.py
@@ -31,7 +31,6 @@ from modulus.distributed.mappings import (
     gather_from_parallel_region,
     scatter_to_parallel_region,
 )
-from modulus.distributed.utils import compute_split_shapes
 from modulus.models.afno.distributed.layers import (
     DistributedAFNO2D,
     DistributedMLP,
@@ -98,9 +97,6 @@ class DistributedBlock(nn.Module):
 
     def forward(self, x):
         if not self.input_is_matmul_parallel:
-            scatter_shapes = compute_split_shapes(
-                x.shape[1], DistributedManager().group_size("model_parallel")
-            )
             x = scatter_to_parallel_region(x, dim=1, group="model_parallel")
 
         residual = x
@@ -117,9 +113,7 @@ class DistributedBlock(nn.Module):
         x = x + residual
 
         if not self.output_is_matmul_parallel:
-            x = gather_from_parallel_region(
-                x, dim=1, shapes=scatter_shapes, group="model_parallel"
-            )
+            x = gather_from_parallel_region(x, dim=1, group="model_parallel")
 
         return x
