+++ modulus/models/diffusion/__init__.py
@@ -25,4 +25,10 @@ from .layers import (
 from .song_unet import SongUNet
 from .dhariwal_unet import DhariwalUNet
 from .unet import UNet
-from .preconditioning import EDMPrecond, VEPrecond, VPPrecond, iDDPMPrecond
+from .preconditioning import (
+    EDMPrecond,
+    EDMPrecondSR,
+    VEPrecond,
+    VPPrecond,
+    iDDPMPrecond,
+)
