+++ modulus/models/gnn_layers/__init__.py
@@ -12,11 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .distributed_graph import (
-    DistributedGraph,
-    GraphPartition,
-    partition_graph_by_coordinate_bbox,
-    partition_graph_nodewise,
-    partition_graph_with_id_mapping,
-)
+from .distributed_graph import DistributedGraph
 from .graph import CuGraphCSC
