+++ modulus/models/gnn_layers/graph.py
@@ -25,7 +25,7 @@ except ImportError:
     # for Python versions < 3.11
     from typing_extensions import Self
 
-from modulus.models.gnn_layers import DistributedGraph, GraphPartition
+from modulus.models.gnn_layers import DistributedGraph
 
 try:
     from pylibcugraphops.pytorch import BipartiteCSC, StaticCSC
@@ -90,7 +90,6 @@ class CuGraphCSC:
         cache_graph: bool = True,
         partition_size: Optional[int] = -1,
         partition_group_name: Optional[str] = None,
-        graph_partition: Optional[GraphPartition] = None,
     ) -> None:
         self.offsets = offsets
         self.indices = indices
@@ -122,7 +121,6 @@ class CuGraphCSC:
             self.indices,
             partition_size,
             partition_group_name,
-            graph_partition=graph_partition,
         )
         # overwrite graph information with local graph after distribution
         self.offsets = self.dist_graph.graph_partition.local_offsets
