+++ modulus/registry/model_registry.py
@@ -12,11 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from importlib.metadata import EntryPoint, entry_points
 from typing import List, Union
 
-# This import is required for compatibility with doctests.
-import importlib_metadata
+import pkg_resources
 
 import modulus
 
@@ -38,7 +36,8 @@ class ModelRegistry:
     @staticmethod
     def _construct_registry() -> dict:
         registry = {}
-        entrypoints = entry_points(group="modulus.models")
+        group = "modulus.models"
+        entrypoints = pkg_resources.iter_entry_points(group)
         for entry_point in entrypoints:
             registry[entry_point.name] = entry_point
         return registry
@@ -100,13 +99,13 @@ class ModelRegistry:
             If no model is registered under the provided name.
         """
 
-        model = self._model_registry.get(name)
-        if model is not None:
-            if isinstance(model, (EntryPoint, importlib_metadata.EntryPoint)):
+        if name in self._model_registry:
+            model = self._model_registry[name]
+            if isinstance(model, pkg_resources.EntryPoint):
                 model = model.load()
             return model
-
-        raise KeyError(f"No model is registered under the name {name}")
+        else:
+            raise KeyError(f"No model is registered under the name {name}")
 
     def list_models(self) -> List[str]:
         """
