+++ modulus/utils/generative/__init__.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from .sampler import ablation_sampler
 from .utils import (
     EasyDict,
     InfiniteSampler,
@@ -21,6 +22,7 @@ from .utils import (
     check_ddp_consistency,
     constant,
     construct_class_by_name,
+    convert_datetime_to_cftime,
     copy_files_and_create_dirs,
     copy_params_and_buffers,
     ddp_sync,
