+++ test/distributed/test_config.py
@@ -135,7 +135,7 @@ def run_distributed_model_config(rank, model_parallel_size, verbose):
 @pytest.mark.multigpu
 def test_distributed_model_config():
     num_gpus = torch.cuda.device_count()
-    assert num_gpus >= 2, "Not enough GPUs available for test"
+    assert num_gpus == 2, "Not enough GPUs available for test"
     model_parallel_size = 2
     verbose = False  # Change to True for debug
