+++ test/distributed/test_distributed_fft.py
@@ -189,7 +189,7 @@ def run_distributed_fft(rank, model_parallel_size, verbose):
 @pytest.mark.multigpu
 def test_distributed_fft():
     num_gpus = torch.cuda.device_count()
-    assert num_gpus >= 2, "Not enough GPUs available for test"
+    assert num_gpus == 2, "Not enough GPUs available for test"
     model_parallel_size = 2
     verbose = False  # Change to True for debug
