+++ test/models/graphcast/test_graphcast_snmg.py
@@ -172,7 +172,7 @@ def run_test_distributed_graphcast(
 
 
 @pytest.mark.multigpu
-@pytest.mark.parametrize("dtype", [torch.float32, torch.float16])
+@pytest.mark.parametrize("dtype", [torch.float32, torch.float16, torch.bfloat16])
 @pytest.mark.parametrize("do_concat_trick", [False, True])
 @pytest.mark.parametrize("do_checkpointing", [False, True])
 def test_distributed_graphcast(dtype, do_concat_trick, do_checkpointing):
